Added support for .bz2 archives as well as .gz.
I came across an issue while basket was downloading pytz.  pytz is a
bz2 archive.  Basket tested to see if the archive was a type that
tarfile could extract, which bz2 is, however in the extraction function
only .gz was being handled.  I simply extracted the extension and then
used an if statement to change which archive format tarfile was
extracting.  Probably can be done a bit more efficient but I'm flying
out in a few moments and wanted to get it uploaded.